Add events package
