+++ common/events/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@casimir/events",
+  "private": "true",
+  "main": "src/index.ts",
+  "scripts": {
+    "build": "echo '@casimir/events build not specified. Disregard this warning and any listed errors above if @casimir/events is not needed for the current project build.' && exit 0",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "dependencies": {
+    "ethers": "^5.7.2"
+  },
+  "devDependencies": {
+    "@types/node": "^17.0.38"
+  }
+}
