+++ common/fetch/package.json
@@ -1,0 +1,12 @@
+{
+    "name": "@casimir/fetch",
+    "private": true,
+    "main": "src/index.ts",
+    "scripts": {
+        "build": "echo '@casimir/fetch build not specified. Disregard this warning and any listed errors above if @casimir/fetch is not needed for the current project build.' && exit 0",
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "devDependencies": {
+        "@types/node": "^17.0.38"
+    }
+}
