+++ common/shell/package.json
@@ -1,0 +1,12 @@
+{
+    "name": "@casimir/shell",
+    "private": true,
+    "main": "src/index.ts",
+    "scripts": {
+        "build": "echo '@casimir/shell build not specified. Disregard this warning and any listed errors above if @casimir/shell is not needed for the current project build.' && exit 0",
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "devDependencies": {
+        "@types/node": "^17.0.38"
+    }
+}
