+++ contracts/ethereum/scripts/test.ts
@@ -1,4 +1,4 @@
-import { run } from '@casimir/helpers'
+import { run } from '@casimir/shell'
 
 /**
  * Test ethereum contracts
