+++ infrastructure/cdk/src/providers/network.ts
@@ -1,9 +1,8 @@
 import { Construct } from 'constructs'
 import * as cdk from 'aws-cdk-lib'
 import * as ec2 from 'aws-cdk-lib/aws-ec2'
-import * as ecs from 'aws-cdk-lib/aws-ecs'
 import { NetworkStackProps } from '../interfaces/StackProps'
-import { pascalCase } from '@casimir/helpers'
+import { pascalCase } from '@casimir/format'
 import { Config } from './config'
 
 /**
