+++ scripts/root/check.ts
@@ -1,4 +1,4 @@
-import { run } from '@casimir/helpers'
+import { run } from '@casimir/shell'
 import fs from 'fs'
 
 /**
