+++ scripts/root/dev.ts
@@ -1,5 +1,5 @@
 import { ethers } from 'ethers'
-import { getWallet, run, runSync } from '@casimir/helpers'
+import { run, runSync } from '@casimir/shell'
 import { loadCredentials, getSecret } from '@casimir/aws'
 
 /**
@@ -20,11 +20,13 @@ void async function () {
 
     if (process.env.USE_SECRETS !== 'false') {
         await loadCredentials()
+        process.env.BIP39_SEED = process.env.BIP39_SEED || await getSecret('consensus-networks-bip39-seed')
+    } else {
+        process.env.BIP39_SEED = process.env.BIP39_SEED || 'inflict ball claim confirm cereal cost note dad mix donate traffic patient'
     }
 
     process.env.PROJECT = process.env.PROJECT || 'casimir'
     process.env.STAGE = process.env.STAGE || 'local'
-    process.env.BIP39_SEED = process.env.USE_SECRETS !== 'false' ? process.env.BIP39_SEED || await getSecret('consensus-networks-bip39-seed') : process.env.BIP39_SEED || 'inflict ball claim confirm cereal cost note dad mix donate traffic patient'
     process.env.CRYPTO_COMPARE_API_KEY = process.env.USE_SECRETS !== 'false' ? process.env.CRYPTO_COMPARE_API_KEY || await getSecret('casimir-crypto-compare-api-key') : process.env.CRYPTO_COMPARE_API_KEY || ''
     process.env.EMULATE = process.env.EMULATE || 'false'
     process.env.FORK = process.env.FORK || 'testnet'
@@ -82,7 +84,7 @@ void async function () {
         const provider = new ethers.providers.JsonRpcProvider(process.env.ETHEREUM_FORK_RPC_URL)
         process.env.ETHEREUM_FORK_BLOCK = process.env.ETHEREUM_FORK_BLOCK || `${await provider.getBlockNumber() - 5}`
 
-        const wallet = getWallet(process.env.BIP39_SEED)
+        const wallet = ethers.Wallet.fromMnemonic(process.env.BIP39_SEED)
 
         // Account for the mock Chainlink functions deployments
         const deployerNonce = await provider.getTransactionCount(wallet.address) + 5
