+++ services/functions/package.json
@@ -1,9 +1,10 @@
 {
     "name": "@casimir/functions",
     "private": true,
-    "main": "index.js",
+    "main": "src/index.ts",
     "scripts": {
         "compile": "hardhat compile",
+        "dev": "SIMULATE_FUNCTIONS=true node index.js",
         "test": "npm run test:unit",
         "test:unit": "hardhat test test/unit/*.spec.js",
         "functions-simulate-javascript": "node scripts/simulateFunctionsJavaScript.js",
