+++ services/oracle/scripts/dev.ts
@@ -1,4 +1,5 @@
-import { fetchRetry, run } from '@casimir/helpers'
+import { fetchRetry } from '@casimir/fetch'
+import { run } from '@casimir/shell'
 
 /**
  * Start development DAO oracle service
