--- services/oracle/src/interfaces/GetEventsIterableInput.ts
@@ -1,8 +1,0 @@
-import { ethers } from 'ethers'
-
-export interface GetEventsIterableInput {
-    ethereumUrl?: string
-    provider?: ethers.providers.JsonRpcProvider
-    managerAddress: string
-    events: string[]
-}
