+++ services/users/scripts/db.ts
@@ -1,6 +1,6 @@
 import fs from 'fs'
 import os from 'os'
-import { run } from '@casimir/helpers'
+import { run } from '@casimir/shell'
 import { getSecret, loadCredentials } from '@casimir/aws'
 import { JsonSchema, Schema, accountSchema, nonceSchema, operatorSchema, userSchema, userAccountSchema } from '@casimir/data'
