+++ services/users/src/providers/db.ts
@@ -1,5 +1,5 @@
 import { Postgres } from './postgres'
-import { camelCase } from '@casimir/helpers'
+import { camelCase } from '@casimir/format'
 import { Account, OperatorAddedSuccess, RemoveAccountOptions, User, UserAddedSuccess, UserWithAccountsAndOperators } from '@casimir/types'
 import useEthers from './ethers'
